Change file processing to async and refactor function name
Refactor from calling direcly convertAndRemoveFile to start separate goroutine for file processing in sink/file.go for non-blocking file operations. This will convert 'filename' from ts to mp4 using 'convertTsToMp4' function then delete the file if there's no error. The function name change makes clearer what its purpose is and gives a better overview for other developers.

+++ sink/file.go
@@ -41,8 +41,13 @@ func NewFileSink(recordCtx record.RecordContext) (chan<- []byte, error) {
 			return
 		}
 
-		_ = convertAndRemoveFile(filename)
-		_ = removeFile(filename)
+		go func() {
+			err := convertTsToMp4(filename)
+			if err == nil {
+
+				_ = removeFile(filename)
+			}
+		}()
 	}()
 
 	return sinkChan, nil
@@ -59,7 +64,7 @@ func isFFmpegInstalled() bool {
 	return true
 }
 
-func convertAndRemoveFile(filename string) error {
+func convertTsToMp4(filename string) error {
 	// Run ffmpeg command to convert .ts to .mp4
 	mp4Filename := fmt.Sprintf("%s.mp4", filename)
 	log.Printf("Converting %s to %s\n", filename, mp4Filename)
